Reservations & File removal

+++ src/main/java/com/location/voiture/DAO/VoitureRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.location.voiture.Entities.Voiture;
 
 public interface VoitureRepository 
-         extends JpaRepository<Voiture, Integer> {
+         extends JpaRepository<Voiture, Long> {
 	
 	@Query("select e from Voiture e where e.disponibilite = 1")	
 	public List<Voiture> disponibiliteVoiture();
